Update SCR.js
Added a new prefix for playing sounds. Rather than !play <filename>, ^<filename> is now the convention. !<cmd> will now be used for text-based functions from now on.
Of course, these prefixes are entirely customizable and are up to the end user. the commandPrefix will be defaulted to '!' and the playPrefix will be defaulted to '^'.

Added more requirements. Now the user is required to not only have node.js, ffmpeg, and discord.js installed, but math.js as well.
math.js was added in order to generate random numbers to be used in playing random sounds.

Made a few functions a bit cleaner, including "play" and "contains."
This is because these functions will be used by different chat commands repeatedly, so it'd be best to keep them easily accessible and less redundant.

Currently, the robot is set to join the user's channel whenever the user requests a sound to be played, regardless of whether or not the robot is already connected. This is redundant and a waste of resources, and I plan on fixing this problem soon. But for now, for the sake of getting things working, I will keep this.

The bot also ignores all messages that do not start with the appropriate prefixes, as well as messages made by other bots to prevent loop spam.

Added the random command. This command is used to play a random sound file in the audioDirectory.

Added a few text commands: index, help, and leave.
each of these commands will send a message to the text channel which the request was received.
index displays all filenames within the audioDirectory.
help displays all ways the user can interact with the robot.
leave will cause the robot to disconnect from its current voice channel.

Currently, 'leave' will also make the robot join the user's channel and then leave. As this is redundant, it will be fixed in a future version.

Also added 'debugMode', a way for the robot to log relevant information to the command prompt. Right now, it will log all commands and sounds played. debugMode is defaulted to 'true'.

Currently, the robot can only join one channel, whose ID must be specified in the code. This will be fixed later, as many servers have multiple channels.

Also, the robot is currently only optimized to be used on ONE server. This will be fixed at a later time, for now I want to focus on getting the robot to work perfectly on a single server.