Update SCR.js
Added a "currentConnection" variable that will store the voiceConnection object. This will eliminate the need for reconnecting to a voice channel every time a sound needs to be played, or every time the robot needs to disconnect from voice.

Added better comments for my methods, "contains" and "play"

Added the "joinChannel" method, so that the robot is no longer limited to joining a single hard-coded channel.

The "join" command will now cause the robot to join the voice channel of the user who made the request. If the user is not in a channel, the robot will return a message reflecting this and do nothing.