Update SCR.js
Cleaned the code up and added more comments.

Added a silence function, which is intended to play a blank sound file, but can play any sound file the user wants. The sound file must be located one directory above the sound file folder, so that the blank file does not appear in the index and is not played upon a 'random' request. User must create this sound file.

Added a shutdown feature, which shuts the robot down gracefully, rather than the user needing to close the command prompt or press ctrl+C.

Improved the if/else hierarchy in the message processing area. Skips more lines depending on the prefix the robot reads.

The 'help' command was updated to reflect the new commands, and was also cleaned up.