Update SCR.js
Learned that the number of voiceConnections and the voiceConnection variable do not update until bot.eval() has finished, so immediately knowing whether or not a join was successful while performing bot.eval() is impossible. Since whether or not the request was successful is obviously apparent within the Discord App, I decided to do away with this feature. The bot will no longer message the user as to whether or not the join to their voice channel was successful. Rather, it will state that it is attempting to join the voice channel as well as when the attempt is complete, and leave it at that.

The bot, however, will still check for errors while connecting and will notify users if there was any problem in that nature regarding connecting.