Changed version name to 1.8
Looking back, I realize that I was quite liberal with version naming.
The original goal for the Soda Can robot was to play audio over voice
using simple commands, and utilizing a simple directory system
consisting of a single folder filled with audio clips, which I will call
"Verison 1." I originally intended this project to be very simple and
barebones, for my use only. The original plan was just for a robot to be
used on a small Discord server amongst close friends.
I will consider Version 2 to be where more advanced folder navigation is
put into play, along with better usability, customization, and
bugfixing. I want Version 2 to be well-optimized and usable for larger
Discord servers, so I intend to add more security features. However,
although I want it to be usable on a large Discord server, I do not plan
on adding moderation features to the robot since that can be
accomplished with different robots that are dedicated to moderation.
That is not what I want for the Soda Can Robot.
Version 3 will include huge changes that I am very excited about, that
will completely change the scope of the Soda Can Robot. Micspamming will
never be the same.